Started pulling in some higher levels of abstraction on time series charts
This is a first pass as this is a repurposed abstract base class that previously was exclusively applied to impulse responses with milliseconds as the units. As a result, in spite of all of those assumptions, and the notion of balancing negative and positive time offsets from zero, have been pushed to the next level up (which hopefully will be published soon as well), there may need to be additional parameterization vs. using current assumptions on desired number of ticks.