Split SPL Palette Legend into a more abstract Palette Legend base class
I had initially assumed that all such Palette Legends would involve calculated dynamic ranges of min/max values and the like, and use standard color palettes such as NASA Jet Palette.

The class split and refactoring was necessitated by a use case where the palette is in the domain of percentiles, whether from 0.0 to 1.0 or 0.0 to 100.0, and where the color shifts may be very specific to domain logic and individual results vs. generalized to a specific domain.