Update PercentPaletteLegend.java
Need to return to axis on the left as there are still some assumptions in the base class about layout order for the axis and the image container box, so setting to the right simply moves the labels to the right of the axis vs. moving the axis to the right of the color box container for the palette image.