A valiant attempt to abstract Impulse Response charting just enough to serve as boilerplate
A reduction of assumptions from downstream application context, but some things hold, based on my observations of several apps dealing with Impulse Response and IFFT. Namely, the y-axis is normalized between { -1, 1 } as the main point of such plots is to find Peak Impulse Time and Impulse Echo Times.

Otherwise, it's a fairly complete implementation with customization for sample rates and such, but the provided default data vector may not be enough for most apps to work with, even if they add more data vectors to the set, so it is implemented as an abstract base class and in such a way that the main data set accessors serve as boilerplate examples.